beta 1.4 --refactor master gui classes as module
